GS Beam Attack Charge-up
Grand Star now "charges up" a beam attack. This spawns a energy circle on the GS that mimics it's movement. It gradually shrinks and a charging laser sound effect plays. The GS will spawn a white puff of smoke indicating it's about to shoot, which is next up to make...