Fixed TTC Thwomp
There was a Thwomp at the top of TTC that guarded the star, or the "flood" flag if that mod is enabled. This Thwomp made collecting either basically impossible. I removed a duplicate function for Thwomps and fixed the check that disabled instant-thwomp crushes if the player is at TTC.

One of the few mercy's I'm willing to grant for Hardmode :D