Added GS Laser Beam
There is now a Laser Beam that shoots out of the grand star, with a sound effect. Collision data is present, but doesn't seem to do anything... Will need to figure this out and add a delay to the beam so it doesn't aim-bot Mario. (Plus add any extra fluff to this)