Added GS Death, various fixes.
The GS can now be hurt, and can also be killed if you ground_pound him when it's stunned.

Fixed the height issues that came up if the GS landed in lava at the end of ACT_JUMP_TOWARDS_MARIO. This fix is in ACT_GO_HOME. The minions will now spawn as soon as GS reaches its target height.