Testing Trophy Collect
Added a test trophy in Hell that is spawned in Blender.. I also added a sound effect and item deletion on the hitbox detection.

The sound is causing errors so you have to comment it out if you want the deletion to work, I'm guessing because we are out of audio_sample slots? (This was sound #26)