Kind of a lot
Some bug fixes, some other things. The beef of this is that Tall Tall Mountain is now DARK DARK MOUNTAIN. It's pitch black.

There is a "lantern" that mario can pickup and take with him. It's currently a bobomb since I was expirimenting with some things. I made a custom lantern behavior that was acting up, so I put it to the side and used bobomb_buddy behavior instead.

After having so much trouble, I decided I will make a much better option for players - A hidden nightvision headset! The programming is all there, but there's no model for it yet. You can equip nightvision by using Dpad Down for the time being. It works well :>

Lots of fog and lighting changes to accomodate Dark Dark Mountain.

Some other fixes for things that popped up here and there. My brain is a scrambled egg and I now go to bed.