TTM Dark/Nightvision
More addons, tweaks, and fixes to Dark Dark Mountain.

Switched bobomb for a carry/throwable lantern that is now working. Added a glow that follows. Added night vision headset and animated it.

The only things left to do now are to model a nightvision headset and then hide it somewhere on the map.